refactor(thirdpartyservicespriveconsts): change type of consts

+++ SalaryService.Domain/Common/ThirdPartyServicesPriceConsts.cs
@@ -3,12 +3,12 @@ namespace SalaryService.Domain.Common
 {
     internal class ThirdPartyServicesPriceConsts
     {
-        public const int AccountingPerYear = 7200;
+        public const double AccountingPerYear = 7200;
 
-        public const int AccountingPerMonth = 600;
+        public const double AccountingPerMonth = 600;
 
-        public const int ParkingCostPerMonth = 1800;
+        public const double ParkingCostPerMonth = 1800;
 
-        public const int ParkingCostPerYear = 21600;
+        public const double ParkingCostPerYear = 21600;
     }
 }
